requirements.txt
appdirs==1.4.3
argon==0.2.0
argon2-cffi==18.1.0
cffi==1.11.2
Django==1.10.2
django-allauth==0.27.0
django-appconf==1.0.2
django-autocomplete-light==3.2.1
django-bootstrap3==9.0.0
django-braces==1.10.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-extra-views==0.8.0
django-markdown-deux==1.0.5
django-model-utils==2.6
django-multiselectfield==0.1.5
django-notifications-hq==1.2
django-rest-framework-braces==0.2.2
django-smartfields==1.0.9
django-tinymce==2.7.0
djangorestframework==3.6.3
humanize==0.5.1
itsdangerous==0.24
markdown2==2.3.4
oauthlib==2.0.6
pep8==1.7.1
Pillow==5.0.0
pycodestyle==2.3.1
pycparser==2.17
pytz==2016.7
requests==2.18.4
requests-oauthlib==0.8.0
selenium==2.53.1
signals==0.0.2
sqlparse==0.2.2
Unidecode==0.4.21
WTForms==2.1
